Complete authorization and registration activities

+++ HSE_Slaves/app/src/main/java/com/hse_project/hse_slaves/model/UserRegistration.kt
@@ -8,7 +8,7 @@ data class UserRegistration(
     var username: String,
     var password: String,
     var specialization: String,
-    var rating: Double,
+    var rating: Int,
     var description: String,
     var images: ArrayList<String>
 )
